Update styling to header and navbar using tailwind
