--- src/components/sections/TopNavigation/TopNavigation.css
@@ -1,13 +1,0 @@
-.top-navigation .search-bar {
-  padding: 0 20px;
-  margin: 0 40px;
-}
-
-.top-navigation li a {
-  padding: 0 10px;
-  text-transform: uppercase;
-}
-
-.top-navigation li a:hover {
-  color: grey;
-}
